fix: repair Windows skill test failures
