+++ understand-anything-plugin/src/__tests__/extract-structure.test.mjs
@@ -1,5 +1,5 @@
 import { describe, it, expect } from "vitest";
-import { buildResult } from "../../skills/understand/extract-structure.mjs";
+import { buildResult } from "../../skills/understand/extract-structure-result.mjs";
 
 const file = (overrides = {}) => ({
   path: "src/foo.py",
